# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.createCommentRequest:
    properties:
      body:
        type: string
    required:
    - body
    type: object
...
  handler.errorResponse:
    properties:
      code:
        type: string
      error:
        type: string
    type: object
...
  handler.updateCommentStatusRequest:
    properties:
      status:
        enum:
        - approved
        - rejected
        type: string
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: A production-ready personal blog REST API
  title: Simple Blog API
  version: "1.0"
paths:
...
  /comments/{id}:
    delete:
      description: Deletes a comment
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete a comment
      tags:
      - comments
  /comments/{id}/status:
    patch:
      consumes:
      - application/json
      description: Approves or rejects a comment
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Status (approved|rejected)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.updateCommentStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Update comment status
      tags:
      - comments
...
  /images:
    post:
      consumes:
      - multipart/form-data
      description: Uploads an image file (max 10MB)
      parameters:
      - description: Image file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Upload an image
      tags:
      - images
  /images/{id}:
    delete:
      description: Deletes an image by ID
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete an image
      tags:
      - images
...
  /posts/{id}/comments:
    get:
      description: Returns paginated comments for a given post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: List comments for a post
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Adds a comment to a post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Comment body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Create a comment
      tags:
      - comments
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"